requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
PyJWT==2.8.0
requests==2.32.3
